renamed widgets with template name
